modify itempoint
1. block define as uint64 for fileid
2. offset define as uint32 for file offset
3. fix all elog itempoint

+++ src/backend/access/gin/ginentrypage.c
@@ -91,7 +91,7 @@ GinFormTuple(GinState *ginstate,
 
 	newsize = SHORTALIGN(newsize);
 
-	GinSetPostingOffset(itup, newsize);
+	GinSetPostingOffset(itup, (uint64)newsize);
 	GinSetNPosting(itup, nipd);
 
 	/*
